docs(user): entity, types 를 aggregate 디렉토리 하위로 move
좀 더 명확한 디렉토리 구조를 구성했습니다.
