+++ user/src/main/java/org/teamone/user/command/domain/aggregate/types/UserStatus.java
@@ -1,4 +1,4 @@
-package org.teamone.user.command.domain.types;
+package org.teamone.user.command.domain.aggregate.types;
 
 public enum UserStatus {
     ACTIVE,
