+++ user/src/main/java/org/teamone/user/command/domain/repository/UserRepository.java
@@ -2,7 +2,7 @@ package org.teamone.user.command.domain.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
-import org.teamone.user.command.domain.entity.User;
+import org.teamone.user.command.domain.aggregate.entity.User;
 
 @Repository
 public interface UserRepository extends JpaRepository<User, Integer> {
